requirements.txt
asgiref==3.4.1
Django==3.2.8
django-widget-tweaks==1.4.9
Faker==8.16.0
coverage==6.0.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
humanize==3.12.0
libgravatar==1.0.0
click==8.0.3
cssselect==1.1.0
django-with-asserts==0.0.1
lxml==4.6.4
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
python-dateutil==2.8.2
gunicorn
django-heroku
